requirements.txt
openai>=1.0.0
langchain>=0.0.270
langchain-openai>=0.0.1
beautifulsoup4>=4.12.2
requests>=2.31.0
sqlite-utils>=3.35
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.1
numpy>=1.24.0
flask>=2.3.0
pytest>=7.4.0
transformers>=4.35.0
tiktoken>=0.5.1
jsonschema>=4.19.1
pytest-cov>=4.1.0
loguru>=0.7.0
rich>=13.6.0 